requirements.txt
openpyxl
python-docx
beautifulsoup4
requests
